Add a caching layer (#2)
Because the corpra are very likely not to change during a session, this
should really not change anything in the package.  It will help in particular
when the package is in a library on a network share, and for my own use
within my ids package.

A possibly better process would be to convert all the .json files into
an object R/sysdata.R and skip the json loading step entirely (this would
need to be done each time the upstream data is updated and not done on
CRAN).  This could reduce the size of the package a little, not sure.